requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
twilio==9.0.4
google-generativeai==0.7.2
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
